LaTeX filke for a template for NEU assignments
A LaTeX file I created for a template that can be used for submission of Northeastern University Assignments. This template is created keeping in mind the templates recommended by the department of Analytics at NEU Toronto . It can , I suppose, with little alteration be adapted to other departments as well.